ddl/requirements.txt
psycopg2-binary
python-decouple